backend/requirements.txt
asgiref==3.7.2
Django==4.2
djangorestframework>=3.14
drf-yasg==1.21.3
djoser==2.2.0
django-cors-headers==4.2.0
drf-extra-fields==3.5.0
flake8==6.0.0
isort==5.12.0
mccabe==0.7.0
pillow==10.0.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==1.0.0
sqlparse==0.4.4
tzdata==2023.3
django-phonenumber-field==7.1.0
